@@ -17,7 +17,8 @@
     "start": "nps",
     "format": "nps format.write",
     "build-vs-code": "ng build vscode --maxWorkers=4",
-    "postinstall": "node ./tools/scripts/postinstall.js"
+    "postinstall": "node ./tools/scripts/postinstall.js",
+    "update": "ng update @nrwl/workspace"
   },
   "build": {
     "appId": "io.nrwl.AngularConsole",
@@ -77,39 +78,42 @@
     ]
   },
   "dependencies": {
-    "@nrwl/nx": "7.8.6",
+    "@nrwl/angular": "8.0.0-rc.1",
+    "@angular/core": "8.0.0-rc.4",
     "ij-rpc-client": "^0.3.2",
     "node-pty-prebuilt": "0.7.6",
     "source-map-explorer": "^1.8.0",
     "stylelint": "^9.10.1",
     "stylelint-config-standard": "^18.2.0",
-    "tmp": "^0.1.0"
+    "tmp": "^0.1.0",
+    "core-js": "2.5.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.9",
-    "@angular-devkit/build-ng-packagr": "0.13.9",
-    "@angular/animations": "7.2.15",
-    "@angular/cdk": "7.3.7",
-    "@angular/cli": "7.3.9",
-    "@angular/common": "7.2.15",
-    "@angular/compiler": "7.2.15",
-    "@angular/compiler-cli": "7.2.15",
-    "@angular/core": "7.2.15",
+    "@nrwl/workspace": "8.0.0-rc.1",
+    "@nrwl/node": "8.0.0-rc.1",
+    "@nrwl/jest": "8.0.0-rc.1",
+    "@nrwl/cypress": "8.0.0-rc.1",
+    "@angular-devkit/build-angular": "0.800.0-rc.4",
+    "@angular-devkit/build-ng-packagr": "0.800.0-rc.4",
+    "@angular/animations": "8.0.0-rc.4",
+    "@angular/cdk": "8.0.0-rc.2",
+    "@angular/cli": "8.0.0-rc.4",
+    "@angular/common": "8.0.0-rc.4",
+    "@angular/compiler": "8.0.0-rc.4",
+    "@angular/compiler-cli": "8.0.0-rc.4",
     "@angular/flex-layout": "7.0.0-beta.24",
-    "@angular/forms": "7.2.15",
-    "@angular/language-service": "7.2.15",
-    "@angular/material": "7.3.7",
-    "@angular/platform-browser": "7.2.15",
-    "@angular/platform-browser-dynamic": "7.2.15",
-    "@angular/router": "7.2.15",
+    "@angular/forms": "8.0.0-rc.4",
+    "@angular/language-service": "8.0.0-rc.4",
+    "@angular/material": "8.0.0-rc.2",
+    "@angular/platform-browser": "8.0.0-rc.4",
+    "@angular/platform-browser-dynamic": "8.0.0-rc.4",
+    "@angular/router": "8.0.0-rc.4",
     "@graphql-modules/logger": "^0.2.3",
     "@nestjs/common": "5.7.0",
     "@nestjs/core": "5.7.0",
     "@nestjs/graphql": "5.5.1",
-    "@nrwl/angular-console-enterprise-electron": "0.3.462",
-    "@nrwl/angular-console-enterprise-frontend": "0.3.462",
-    "@nrwl/builders": "7.8.6",
-    "@nrwl/schematics": "7.8.6",
+    "@nrwl/angular-console-enterprise-electron": "0.3.507",
+    "@nrwl/angular-console-enterprise-frontend": "0.3.507",
     "@schematics/angular": "7.3.1",
     "@types/d3-hierarchy": "^1.1.6",
     "@types/d3-selection": "^1.4.1",
@@ -166,11 +170,11 @@
     "graphql-tag": "2.10.1",
     "graphql-tools": "4.0.4",
     "husky": "^1.3.1",
-    "jasmine-marbles": "0.4.0",
+    "jasmine-marbles": "0.5.0",
     "jest": "24.1.0",
     "jest-jasmine2": "23.6.0",
     "jest-preset-angular": "7.0.0",
-    "ng-packagr": "4.7.0",
+    "ng-packagr": "5.2.0",
     "node-pty-prebuilt": "0.7.6",
     "nps": "5.9.3",
     "nps-utils": "1.7.0",
@@ -192,14 +196,17 @@
     "tslib": "1.9.3",
     "tslint": "5.12.1",
     "tslint-no-unused-expression-chai": "^0.1.4",
-    "typescript": "3.2.4",
+    "typescript": "3.4.5",
     "universal-analytics": "0.4.20",
     "uuid": "3.3.2",
     "vsce": "^1.62.0",
     "vscode": "^1.1.33",
     "xterm": "3.11.0",
     "zlib": "^1.0.5",
-    "zone.js": "0.8.29"
+    "zone.js": "0.9.1"
+  },
+  "resolutions": {
+    "webpack": "4.29.0"
   },
   "jest": {
     "moduleFileExtensions": [
